feat(generator): add generator functions for action and import types to streamline code generation process refactor(gen-output-file): simplify by removing inline implementations of genAction and genImportTypes, improving modularity and readability
